chore: 초기세팅 (code-base UI)
